[Tempest] Fix visibility of test_quotas.py module
Commit [1] added wrong decorator to test class located in
"manila_tempest_tests/tests/api/test_quotas.py" module that
made it invisible for tempest.

It should be "ddt.ddt" instead of "ddt.data" as it is now.

Also, fix negative quota tests that were testing wrong thing.
It should have been testing "quota-sets" and "os-quota-sets", not
"services" and "os-services" as it is now.

[1] I82f00114db985b4b3bf4db0a64191559508ac600

Change-Id: Ie0eb7d32b7b032ffdb7f7dd47f68841211e7d7a6
Closes-Bug: #1635588