Fix microversion request URL
The code in the client assumes that we're hitting the
base URL. However, the manila endpoint URL isn't always
of the format HOST[:PORT]/VERSION/xyzzy. It's possible
that the endpoint URL pertains to a proxy configuration
which is of the format
HOST/PROXY/ENDPOINT/FOR/MANILA/VERSION/xyzzy.

This patch also adds logging to the microversion
request since we're using tempest's "raw_request"
which doesn't do logging for us.

Change-Id: I50906ec245b965faaba630fa4d01a8137d4bc3d5
Partial-Bug: #1815038