Add devstack plugin to ease integration tests
Instead of requiring users to install the plugin manually, add a
devstack plugin that can be enabled in order to perform the
installation.

This change should allow us to remove the tempest plugin installation
code from manila's devstack plugin and enable it in job definitions
the same way we enable plugins for manila itself and for the manila-ui
plugin.

Change-Id: If0e5daab19ec83e2da5709693d18fcee72b42766