Fix CIFS access for non-multitenant setup
In the case that multitenancy_enabled is false, the CIFS tests fail,
first because the auxiliary access function returns None, which means
the access object cannot be used by the caller, and second because if a
floating IP address isn't chosen explicitly as the share point, the
server's fixed IP may be chosen which is not in the same segment as the
share server, which means the share will not be accessible. This patch
fixes both issues by making the auxiliary access granting function
consistent for both NFS and CIFS and differentiating instead on
multitenancy_enabled.

Change-Id: Id6a44ea8c6b88d121983566dbb176f48e470094f