Add knowledge of storage network to scenario tests
In many deployments, shared file systems are exported
over a shared storage network. Typically, this shared
storage network is not reachable from within
private tenant networks. Our tests currently
have no cognizance of this topology, and expect that
shares are always exported on a network routable or
directly reachable through private tenant networks.

Add a configuration option to allow configuring the
name or UUID of a shared storage network. If specified
this network will be plugged into VMs created by the
scenario tests and access control will be on the
storage network IP, if NFS.

Change-Id: I8d8063d46284076499d868a5f9b3d13a2a78b5b0
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>