Zuulv3 native "standalone" GlusterFS native, NFS jobs
GlusterFS jobs were included in this repository as
non-voting experimental jobs. However, they depend
on devstack-gate, a deprecated component. Converting
them to native zuulv3 style will allow ease of
maintenance. These jobs were only ever running "api"
tests and were not testing integration with the rest
of OpenStack, so to conserve resources, these jobs
deploy a minimal devstack with just keystone and manila
enabled.

The GlusterFS-NFS job was fixed in the Train release,
we can promote it to the regular CI check queue,
keeping it non-voting at the moment.

However, the GlusterFS-native job has been pretty
failure prone. The legacy DSVM job has no history of
succeeding as of this commit [1]. We'll leave it
in the experimental queue until we triage what's
wrong with the job.

While NFS and Native testing was refreshed in the
Train release, "heketi" layout remains an unmaintained
component of the GlusterFS driver for Manila. This
commit deletes the experimental jobs for heketi.
These can be revived when a maintainer shows up,
willing to invest in the heketi layout for GlusterFS.

[1] https://zuul.opendev.org/t/openstack/builds?job_name=manila-tempest-dsvm-glusterfs-native
Change-Id: Iaef7873e7539045753bb5676fe93f4ded586b5ac
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>