Stop forcing creation of "isolated" credentials
A number of tests are creating an "isolated"
test credential to test cross project separation.
This isn't necessary, since test base classes
already request projects to run the tests. So we
can just use different project contexts where
appropriate.

In the quota tests, we request tempest to allocate
us fresh credentials, because these tests can be
destructive if they accidentally malfunction.
We do this by setting the class level directive
"force_tenant_isolation=True". Whenever we do
this, we will skip the tests if tempest is
configured to not use dynamic credentials;
because tempest will gladly ignore the explicit
configuration to disallow dynamic credentials
and give us new credentials, which might be
frustrating to test runners.

In this change, we
- delete all the code to request "isolated"
  test credentials from tempest.
- delete code that was written to cleanup
  test credentials allocated to test "methods"
  and classes.
- refactor quota tests to revert updates
  when the test concludes (requesting fresh
  credentials for these tests is just a fail-safe)
- Fix creation of share types per visibility, a change
  that was accidentally removed in [1]
- Fix "SharesAdminQuotasUpdateTest.test_reset_share_type_quotas"
  by using a correct project user instead of a
  random project's user to force enforce user quota
  higher than the project.

[1] I15880e400df30918762ebd7304244b4a27200168
Co-Authored-By: Lucio Seki <luciomitsuru.seki@fit-tecnologia.org.br>
Change-Id: I7607019e47ce9e8c6584f69c2b2d3228873f24df
Closes-Bug: #1883795
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>