[Negative] Create share with non allowed share type
This test was created due to a bug [1] that was found when
creating a share type with non allowed share type.
When creating a share using a private share from other tenant, i.e.
the tenant has no access to the share type. As expected it failed
with 404, but the share stuck on "creating" status.

Note: This issue happens only when the share type name is used
for creating the share.
Creating a share using share type id works as expected.

[1] https://bugs.launchpad.net/manila/+bug/1885956

Depends-On: https://review.opendev.org/#/c/738941/
Change-Id: I567abccdaa622193b44a9b70bc0481842d8d55f1
Related-Bug: #1885956