Ping to export location until it is reachable
In some cases it takes time for the interface to acquire the
IP address. This causes a failure in the ping operation.
Therefore, we should ping until the network is reachable.
This is can be done by "call_until_true" method that allows
reusing the ping method until the defined timeout.

This patch also adds the other scenario tests for ipv6
environmet.

Change-Id: I60bf3bd1bd2cc22b5bb6be11c8174b7529d0a066