Fix skip condition for "test_default_share_group_type_applied"
Test "test_default_share_group_type_applied" is failed with code 404
that because there is a wrong skip condition. The test runs when
"default_share_type_name" configuration option is defined but that
does not mean that group share type also exists.
So in case there is a default share type but not a group share type
the api call for default share group types fails because resource
could not be found.
This patch skips the test in case group share type is not found.

Change-Id: I7e16b341e3c9d76647eb2fa08909392328fb330b