SSH over IPv4 when using a storage network
A storage network can be configured to connect
VMs in scenario tests to their shares. When
using a storage network, we don't need to force
the private network that we use to SSH to the
VM to be on a particular IP version.

We need the private network subnet to be of
a particular IP version only when there is
no storage network, and we're connecting to
shares over the same network as SSH.

Change-Id: If1d581730c42b060e74600ef2a05298921b1a3c6
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>