Use the same access rule for both instances
In the test "test_read_write_two_vms" 2 instances are created and 1 share.
To gain access to the share, an access rule is created for each instance.
In cephfs native we don't really need to create a new access rule for
the second instance.
We can use the first access rule for both instances and only create the
configuration files in the instance according to the user account.

Change-Id: I1a15728d47c3fbb1759b6ca44be4ecb25f73ab2c