Tests: reduce concurrency for cephfs-nfs job
Reduce it to 2, which should ensure that this job does
not fail for lack of resources in devstack gate.  We can
bump it back up later if we have test infrastructure improvements,
but in the mean time there is no point holding up fixes and
enhancements to Manila itself because of vulnerabilities in
the test framework itself.

Change-Id: Icf18d737859171510b8f7db6e088d63041fd0bf4