Fix extra specs for share type
In case the backends don't have the same capabilities
share creation can fail. For instance, if one of the
backend has the ability to create share from snapshot
and the other does not.
In case "capability_create_share_from_snapshot_support"
parameter is defined in tempest.conf and it is also
specified in extra specs one of the share will fail to
create.

Change-Id: Iade5b2658df59b7d6cdc9f69418afb16a0108353