Add "member" role to admin project member users
In some tests where we'd like to test role based
access control, and specifically the impact of user
roles (admin vs member vs reader), we create a user
on the fly and assign the necessary roles to this
user.

Recently devstack stopped setting a tempest conf
option that gave us a default role to assign to
regular non-privileged "member" users. [1].

This made the tests that used such users fail
with "401 Unauthenticated" errors.

So we'll need to explicitly request this regular
user role (aka "member") even when not configured
in tempest.conf.

[1] https://review.opendev.org/774524

Change-Id: I103ccda367123bc2429d1d840c0201b01ac61b4d
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>