Add extra check to server migration complete test
The change [1] modified the share server migration complete
operation behavior in Manila.

In the previous behavior there was a chance that manila would
keep the inactive source share server after a successful
migration.

Such behavior is not possible anymore after this change was
merged, since Manila started to expect that the source share
servers were deleted within the successful completion of a share
server migration.

This change introduces a new waiter to make sure that the source
share server is actually being deleted from manila after the
migration complete phase.

[1] https://review.opendev.org/c/openstack/manila/+/803623

Change-Id: Ibb092111032241a71763da0c7fa40470a3cd95ae