Remove share metadata test
Removes test_try_delete_metadata_with_unsupport_format_key
for metadata api refactoring introduced in [1].

This test relates to this bug [2], but fails in the code refactoring
mentioned above. Now that the router handles metadata endpoints, the
stripping of '.html' or '.xml' in the key is unecessary.

For more dicussion see [3].

[1] https://review.opendev.org/c/openstack/manila/+/824648
[2] https://bugs.launchpad.net/manila/+bug/1774353
[3]
https://meetings.opendev.org/meetings/manila/2022/manila.2022-02-03-15.00.log.html#l-77

Change-Id: I3e7d3a49cec6af347368757f30d4478841dabeaa