Follow up fixes to recycle bin tests
For the sake of correctness and readability, the negative
tests to soft-delete or restore with an incorrect ID must
use the default ("primary" user) client in the class. While
it's okay to use whatever client, someone reading these tests
may be confused if the tests are supposed to work with the
default client.

TrivialFix

Change-Id: Ic2735cb3a852af6f8a2edf44a501ba612391ae24
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>