[CI] Add cephadm based native cephfs test job
Add a new job for testing manila's Native CephFS driver
with a cephadm deployed Ceph cluster. This job inherits
from the existing "base" job and overrides the
ceph deployment strategy to use cephadm.

With this commit, we will use this new cephadm based job
instead of the package based deployment to test this
repository. As with the existing job, this new job is
non-voting as well. If the job's stable for a period of
time, we can decide to promote it to being "voting".

Once all CI consumers adopt the new cephadm job, we'll
rewrite the "base" job definition to only care about
cephadm based deployment.

Change-Id: I6e4db6b89f89d2374580827e2334635dc5ead5b9
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>