Split generic job by tests and reduce concurrency
We have been performing some tests using the Ubuntu Jammy image.
While doing such tests, we realized that the newest image would
need more RAM to be spawned successfuly, and that brings the need
to bump the memory for such VMs. While testing with the memory
bump, we realized that the generic driver would face issues with
such change.

We have seen a couple of tests failing due to the share manager
service loosing connection to the database, and this is usually
caused when we have a lack or high consumiption of resources in
a VM. Reducing the concurrency itself should work, and then
we would have the jobs passing. However, in reducing the
concurrency, this job will take longer to run, considering it
will only use one of the available threads, and it can lead to
this job's possibility to hit the 3 hours pre-scheduled limit.

To avoid that, we decided to:

Have this change to reduce the concurrency of the manila generic
driver job to 1, so we can avoid these failures, as well as
split the generic driver job into two jobs:

- manila-tempest-plugin-generic
- manila-tempest-plugin-generic-scenario

The generic one will take care of the API testing as we currently
know and have been doing. The scenario one will run all of the
scenario tests.

Change-Id: Id2d35ce592c73f0c731deb17a5d574f502fe64cb