Pin Ubuntu Focal to manila-service-image on stable branches
Due to some test constraints, we decided to migrate the
manila-service-image to Jammy Jellyfish [1]. In order to have the
jobs passing, we needed to grant more resources to the service VMs,
but it would not be ideal to backport those changes to the stable
branches.

We decided we would pin the manila-service-image to a ubuntu Focal
image in the stable branch, and with that we would not need to
backport the other changes we merged in order to switch the image.

[1] https://review.opendev.org/c/openstack/manila-image-elements/+/841264

Change-Id: I74f81e92297ad58ebd4266511bb62b5fe24f7cb5