Change the set of jobs in the stable branches
In order to switch our manila service image to Ubuntu Jammy, we
needed to get in a couple of changes that Modified the flavor for
the manila service image. After all changes got merged, we switched
to the Ubuntu Jammy image [1].

However, now the manila service image master on Tarballs is using
a Ubuntu Jammy image, which requires more resources. It would not
be ideal to backport the changes that modify the flavor to the
stable branches because it is not an actual requirement for them,
and that also forced us to reduce concurrency of some jobs on the
master branch.

Using the Ubuntu Jammy image broke our stable branches CI, since
with less resources we are not able to successfuly spawn a nova
instance using the Ubuntu Jammy image.

In the stable branches, the jobs that spawn VMs will need to use
Ubuntu focal, so we have a new couple of jobs that will
selectively run against the stable branches and they will use a
pinned version of Ubuntu Focal from Tarballs.

These jobs were grouped into project templates to help us keep our
manifests readable.

Change-Id: I14afe2572c5bc8064417ba7ba3be26751022901d