Use ddt.named_data decorator to avoid random test names
There are few tests that generate random test names because
random parameters are given to the ddt.data  decorator.

In the ddt version number 1.6.0 there is a possibility to
determine the name of the tests without relying on the
given values. I think this could be a good solution for
the randomly generated names.

This change is being re-introduced as ddt version 1.6.0 is
now usable in the downstream CI. The original change is: Iaf39f4a83e383fb9db7f519136b9169352ff84af

The following change is also included: I4d274869e5e9e7800be44bccf7a2641b4a5febe8
Which replaces periods to underscores, a needed fix.

Change-Id: Iaf39f4a83e383fb9db7f519136b9169352ff84ad