Temporary pin the Manila jobs nodeset to Focal
As per the community wide goal to migrate the CI/CD from
Ubuntu Focal to Ubuntu jammy, devstack change was merged, image got
migrated on Nov 18. But Manila jobs are failing on Ubuntu Jammy.

To unblock the Manila CI we need to pin the jobs to ubuntu focal,
until all the jobs are fixed.

Change-Id: Ia235cf28382947a392151f2914ce2dda964b46a6