Correct share group tests cleanup order
Some share group tests were taking advantage of a bug in Manila.
Share network subnets were being allowed to be deleted even when
its share servers were tied to a share group. With that bug being
fixed, we need to ensure that the share groups are deleted before
the share network subnet alongside the share servers are.

Related-Bug: #2004212
Change-Id: I05607883c34bca586c795c7efab1df8d0edd6cff