[CI] Ceph/NFS: skip data-intensive tests with ipv6
We're running test_create_shrink_and_write
and test_create_extend_and_write tests twice,
changing only the client VM's IP address family.
Not much value in running all of these since we're
exhausting the disk space with these tests. We can
choose to re-enable these tests when the situation
changes with either the footprint of the tests
or the form factor of the devstack node.

Partial-Bug: #2009083
Change-Id: I868b2e40934e5d10eb0c5cc6fa867e6adc2cc9fc
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>