[RBAC] Wait for access rules to be applied prior to deletion
For some backends, the access rules might take some time to apply,
and if we try to clean up the rule before it was applied, we would
end up in a race condition.

In order to avoid this issue, we can wait for the access rule to be
active before proceeding to do other steps in the test.

Change-Id: Ib57eda56fdbec398d4dc6e80e982ace5f14727bb