Add manila.conf parameters for NetApp Active IQ weigher
As suggested in a manila meeting [1], would be good have the NetApp Active IQ weigher configured in the dummy jobs to make sure this new weigher
will not break anything. The expected behavior is that the weigher is skipped, since are non-NetApp backends configured. For more information
about this weigher check the Active IQ weigher spec [2].

This patch change the dummy-dhss and dummy-no-dhss jobs to test this scenario by adding the Active IQ necessary parameters in the manila.conf.

Depends-On: I36b08066545afdaa37e053eee319bc9cd489efdc

[1] https://meetings.opendev.org/meetings/manila/2023/manila.2023-08-10-15.00.log.html

[2] https://specs.openstack.org/openstack/manila-specs/specs/bobcat/spec-lite-netapp-activeiq-weigher.html

Change-Id: Ie79404e476ca5bb29c5bfb2f3235559969b8cb55