Use Ubuntu Jammy and Temporarily disable scenario tests
A couple of issues happening at the moment:

- Nova recently merged a change that bumps the libvirt version in
the Bobcat release [1]. The new libvirt version is not available
in Ubuntu Focal. In order to get the new libvirt version, our jobs
that were using Ubuntu Focal, are now forced to use Jammy, so we
bump it to the new version in this patch.

- The support for Ubuntu Focal has also been dropped in this
release [2]. Making it one more reason to get the new jammy
images.

While upgrading the image, some other issues surfaced:

- Some CI jobs are doing IPv6 testing, which use quagga. Quagga
does not have a candidate for installation in Jammy and must be
replaced. This issue is already documented [3].

- The inability to set up IPv6, combined with the scenario tests
being enabled in the LVM and CEPH drivers brought up another known
issue: when doing IPv6 setup, we don't set up only IPv6 networking
and we also end up setting crucial information for the scenario
tests to run in the function that does IPv6 setup. So in case the
job is only using IPv4 in scenario testing, we currently don't set
some key configuration to let the shares be created/mounted [4].

- Due to the reasons above, the scenario tests will never pass, as
we can't use Focal images, and when using Jammy, we can't do
scenario testing.

This is blocking the Manila gate and all other changes in the CI.
As we are forced to use the Jammy image and the scenario tests
won't pass, we disable them in this change, do some refactoring
in Manila [5], and re-enable them again for IPv4, until we have
figured out the quagga replacement.

[1] https://review.opendev.org/c/openstack/nova/+/887255
[2] https://lists.openstack.org/pipermail/openstack-discuss/2023-September/034968.html
[3] https://bugs.launchpad.net/manila/+bug/1998489
[4] https://github.com/openstack/manila/blob/master/devstack/plugin.sh#L1104-L1116
[5] https://review.opendev.org/c/openstack/manila/+/893804

Change-Id: Ia45a8e7f7c9cf9959bcc9b9a7243cccbbd1f4916