[CI] Stable branch CI adjustments
Xena and Yoga are past the stable maintenance
cycles. We're soon going to tag these release
branches as "unmaintained" branches [1]. We don't
expect the trunk version of manila-tempest-plugin
to work against these older branches any more.
Instead, we have tagged a "xena-last" and will
soon [2] tag a "yoga-last" branch that will be more
appropriate.

Going forward we must be testing the "maintained"
stable branches, which are zed, antelope, bobcat
as of this commit.

[1] https://docs.openstack.org/project-team-guide/stable-branches.html#unmaintained
[2] https://review.opendev.org/c/openstack/releases/+/908381
Change-Id: Ib0e984a4acdc2f82f5e3a75e34c7ec922674b0a5
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>