Add CEPHFS filesystem metadata verifications
Recently, we have implemented a new feature that allows the share
drivers to set metadata to the shares on the creation [1].

This feature allowed the Ceph driver to report the filesystem name
where the share was created (only when using the CEPHFS protocol),
making it easier for users to get this data before they mount
their shares.

This change adds a tempest test that will check if such config
is present in the shares.

[1] https://review.opendev.org/c/openstack/manila/+/909175
[2] https://review.opendev.org/c/openstack/manila/+/909916

Change-Id: I2ecef30de76c40b98accb5ce04b6c9eefdf7fea2
Signed-off-by: Carlos Eduardo <ces.eduardo98@gmail.com>