Fix floating IP association by filtering port list using IP address
The recent change in Tempest introduced a step to ensure all VM ports are active before associating a floating IP.
This caused issues in the manila-tempest-plugin scenario tests, where the correct port couldn't be identified for floating IP association.

This patch fixes the issue by filtering the port list using the IP address of the fixed IP to narrow down the result to the specific port associated with the given address.

Depends-On: https://review.opendev.org/c/openstack/tempest/+/937406
Related-Bug: #2089798
Change-Id: I74ac4eafa41f0d6355716e7fe26d09b7dfcbc982