Drop explicit dependency on python-subunit
The library has never been used directly from any code in this
repository. It was used globally before migration to stestr but this
repository has been using stestr since initial project bootstrap.

Note that stestr is still installed as a dependency of stestr.

See [1] and [2] from manila as a reference.

[1] e8bca58887880b53ab61810a414ba8d5c5643c5c
[2] 8ec49f017286fa63b26401f7f3f9a6d45c1a487a

Change-Id: Ib0dba08a05c4f6ba3294ebf17d0298e1d12c62bd
Signed-off-by: Takashi Kajinami <kajinamit@oss.nttdata.com>