Stop testing the V1 API
The /v1 API has been deprecated for a long time [1]; the tempest plugin
will stop testing it so it can be removed in the Manila service.

Removal of this code will affect testing in older maintained stable
branches; however, the long deprecation is reason enough to not go
lengths to preserve testing in these older branches.

[1] https://lists.openstack.org/pipermail/openstack-dev/2016-June/096517.html

Change-Id: Id90044bca44387864aee8984bc2366c089c87cfc
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>